requirements.txt
fastapi
uvicorn[standard]
youtube-transcript-api
yt-dlp
openai-whisper
torch
numpy
python-multipart
requests
